# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MongoVision API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /connections/{connectionId}/databases/{dbName}/collections/{collectionName}/schema:
    get:
      operationId: analyzeSchema
      tags: [schema]
      summary: Analyze and infer schema for a collection
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: path
          required: true
          schema:
            type: string
        - name: sampleSize
          in: query
          schema:
            type: number
            default: 100
      responses:
        "200":
          description: Schema analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaAnalysis"
...
components:
  schemas:
...
    SchemaField:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
        types:
          type: array
          items:
            type: string
        prevalence:
          type: number
        nullable:
          type: boolean
        isArray:
          type: boolean
        isNested:
          type: boolean
        children:
          type: array
          items:
            $ref: "#/components/schemas/SchemaField"
        sampleValues:
          type: array
          items: {}
      required: [name, path, type, prevalence]
    SchemaAnalysis:
      type: object
      properties:
        collectionName:
          type: string
        documentCount:
          type: number
        sampleSize:
          type: number
        fields:
          type: array
          items:
            $ref: "#/components/schemas/SchemaField"
        inconsistencies:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              issue:
                type: string
              affectedDocuments:
                type: number
            required: [field, issue]
      required: [collectionName, documentCount, sampleSize, fields, inconsistencies]
...