# lib/api-zod/src/generated/types/schemaAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MongoVision API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SchemaAnalysisInconsistenciesItem } from "./schemaAnalysisInconsistenciesItem";
import type { SchemaField } from "./schemaField";

export interface SchemaAnalysis {
  collectionName: string;
  documentCount: number;
  sampleSize: number;
  fields: SchemaField[];
  inconsistencies: SchemaAnalysisInconsistenciesItem[];
}